+++ include/perft.h
@@ -4,4 +4,4 @@ uint64_t perftMovegen(Board b, const int depth, const int divide);
 int hashPerft(Board b, const int depth, const uint64_t prevHash);
 void initDummy(void);
 void freeDummy(void);
-int nnuePerft(Board b, const int depth, int32_t* test);
+int nnuePerft(Board b, const int depth, int16_t* test);
